feat: added base HC3 card

+++ lib/entities/card_icon.dart
@@ -1,7 +1,7 @@
 class CardGraphic {
   final String url;
   final String imageType;
-  final int aspectRatio;
+  final double aspectRatio;
 
   CardGraphic({
     required this.url,
@@ -13,7 +13,7 @@ class CardGraphic {
     return CardGraphic(
       url: json['image_url'] as String,
       imageType: json['image_type'] as String,
-      aspectRatio: json['aspect_ratio'] as int,
+      aspectRatio: json['aspect_ratio'] as double,
     );
   }
 }
